fix test timeout (#158)
A glance at the
[actions](https://github.com/acquire-project/acquire-driver-zarr/actions)
lately shows that all testing is timing out on Windows only, and on
`test-get-meta` only. I can't reproduce this on my machine, but I
suspect this is because threads are being shut down before they have a
chance to properly start, since we're not actually doing anything with
them in that test, so they're getting left in a bad state where they're
joinable but we can't join them. So instead of spinning up threads on
the thread pool's construction, I do it in `Zarr::start`. That seems to
fix the problem.