Support multiscale arrays in Zarr V3 with provisional metadata (#269)
This PR uses the same OME metadata as our Zarr V2 writer to express
multiscale semantics. The following changes were made:

- A protected method, `Zarr::make_multiscale_metadata_()`, was
implemented by moving the code from `ZarrV2::write_group_metadata_()` so
that it could be made accessible to both `ZarrV2` and `ZarrV3`
instances.
- Call `make_multiscale_metadata_()` from both overrides of
`write_group_metadata_()`.
- Update the metadata returned by `Zarr::get_meta()` to reflect that
both types of Zarr store support multiscale.

Basic functionality is verified in the
`write-zarr-v3-raw-multiscale.cpp` test file. The test scheme is the
same as in `write-zarr-v2-raw-multiscale.cpp`, but the paths are changed
to reflect the different layout.