Write array metadata whenever array data is written (#282)
- Changes the name of `Writer` (and `ZarrV2Writer` and `ZarrV3Writer`)
class to `ArrayWriter` (resp. `ZarrV2ArrayWriter` and
`ZarrV3ArrayWriter`) to clarify that these classes are responsible for
writing individual arrays (i.e., levels of detail).
- Removes the responsibility for writing array metadata from the `Zarr`
class (and its subclasses) and puts it on the `ArrayWriter` class (and
its subclasses).
- Other metadata (i.e., base, external, and group metadata) are still
the responsibility of the `Zarr` class.
- Sinks to which these other metadata are written are now stored in an
associative array (`std::unordered_map`) rather than a positional array
(`std::vector`) as previously.
- Adds a method `make_sink()` to `SinkCreator` to allow for creation of
array metadata sinks.

This should be the last change before the next release (0.1.12).